// desktop/src/shared/platform-contract-types.ts
export type HanaPlatformInvokeMethodName =
  | "getServerPort"
  | "getServerToken"
  | "runEditCommand"
  | "getAppVersion"
  | "openSettings"
  | "openBrowserViewer"
  | "getVscodeWorkbenchStatus"
  | "openVscodeWorkbench"
  | "closeVscodeWorkbench"
  | "reloadVscodeWorkbench"
  | "setVscodeWorkbenchBounds"
  | "browserGoBack"
  | "browserGoForward"
  | "browserReload"
  | "browserNewTab"
  | "browserSwitchTab"
  | "browserCloseTab"
  | "closeBrowserViewer"
  | "browserEmergencyStop"
  | "selectFolder"
  | "selectFiles"
  | "selectSkill"
  | "selectPlugin"
  | "readFile"
  | "writeFile"
  | "writeFileBinary"
  | "copyFile"
  | "readFileSnapshot"
  | "writeFileIfUnchanged"
  | "watchFile"
  | "unwatchFile"
  | "watchWorkspace"
  | "unwatchWorkspace"
  | "readFileBase64"
  | "readDocxHtml"
  | "readXlsxHtml"
  | "spawnViewer"
  | "viewerClose"
  | "openFolder"
  | "openFile"
  | "openExternal"
  | "showInFinder"
  | "trashItem"
  | "openSkillViewer"
  | "listSkillFiles"
  | "readSkillFile"
  | "closeSkillViewer"
  | "getAvatarPath"
  | "getSplashInfo"
  | "reloadMainWindow"
  | "onboardingComplete"
  | "debugOpenOnboarding"
  | "debugOpenOnboardingPreview"
  | "showNotification"
  | "getAutoLaunchStatus"
  | "setAutoLaunchEnabled"
  | "getKeepAwakeStatus"
  | "setKeepAwakeEnabled"
  | "quickChatReloadShortcut"
  | "quickChatShortcutStatus"
  | "quickChatShow"
  | "quickChatHide"
  | "quickChatResize"
  | "quickChatOpenSession"
  | "getPlatform"
  | "windowMinimize"
  | "windowMaximize"
  | "windowClose"
  | "windowIsMaximized"
  | "appReady"
  | "screenshotRender";

export type HanaPlatformSendMethodName =
  | "syncWindowTheme"
  | "settingsChanged"
  | "startDrag";

export type HanaPlatformEventMethodName =
  | "onQuickChatOpenSession"
  | "onQuickChatShown"
  | "onFileChanged"
  | "onWorkspaceChanged"
  | "onShowSkillViewer"
  | "onSettingsChanged"
  | "onOpenSettingsModal"
  | "onSwitchTab"
  | "onServerRestarted"
  | "onBrowserUpdate"
  | "onVscodeWorkbenchStatus"
  | "onViewerLoad"
  | "onViewerClosed"
  | "onSkillViewerLoad";

export type HanaPlatformFallbackMethodName =
  | "getFileUrl"
  | "getFilePath"
  | "closeBrowser"
  | "onMaximizeChange";

export type HanaPlatformApiMethodName =
  | HanaPlatformInvokeMethodName
  | HanaPlatformSendMethodName
  | HanaPlatformEventMethodName
  | HanaPlatformFallbackMethodName;
// ...
export type HanaPlatformMobileFallbackMethodName = Extract<
  HanaPlatformApiMethodName,
  | "getServerPort"
  | "getServerToken"
  | "openSettings"
  | "onOpenSettingsModal"
  | "getFileUrl"
  | "openFile"
  | "openExternal"
  | "settingsChanged"
  | "syncWindowTheme"
  | "onSettingsChanged"
  | "onSwitchTab"
  | "onServerRestarted"
  | "appReady"
  | "getPlatform"
  | "showNotification"
  | "getAppVersion"
  | "getVscodeWorkbenchStatus"
>;